Squashed commit of the following:
-bumped version to 0.10.93

commit dfa24f68c4cb0e84ad37ac663ec08caadd2f3fbb
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 11:53:32 2022 +0200

    #883 - Remove unused selection action

commit 0ae63081e822a514443a7b6c6e1fb3826a8c6178
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 11:43:17 2022 +0200

    #883 - Remove countOfPendingVectorLoadRequests from selectionReducers with unused hook

commit b884e0b59c5e9000422adbf6e6adf7b14847b51f
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 11:10:44 2022 +0200

    #883 - Remove associatedDownloadTagName from selectionReducers

commit ab825ccf3369af8fbaab9c8d25d68c5a1699d773
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 10:54:57 2022 +0200

    #883 - Remove displayedMoleculesInHitNav from selectionReducers

commit 4103e3984bc9b3fbf2cc48bef3b9d65864a4814b
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 10:08:28 2022 +0200

    #883 - Remove displayAllInNGLList from selectionReducers

commit f3b3b10a879447705b2709b88d48b62fe50e14cc
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 09:59:23 2022 +0200

    #883 - Remove listAllList from selectionReducers

commit 02aa5955918abace2d5473e89bb9260ff10d7338
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 09:40:29 2022 +0200

    #883 - Move categoryList from selectionReducers to apiReducers

commit d059ee8d4de0ddc34be64ef9c55a71c4f2bd96df
Author: Anton Gasperak <anton.gasperak@m2ms.sk>
Date:   Wed Jun 1 08:43:26 2022 +0200

    #883 - Move tagList from selectionReducers to apiReducers
