+++ js/components/preview/molecule/moleculeView/moleculeView.js
@@ -279,7 +279,7 @@ const MoleculeView = memo(
     const [img_data, setImg_data] = useState(img_data_init);
 
     const viewParams = useSelector(state => state.nglReducers.viewParams);
-    const tagList = useSelector(state => state.selectionReducers.tagList);
+    const tagList = useSelector(state => state.apiReducers.tagList);
 
     const { getNglView } = useContext(NglContext);
     const stage = getNglView(VIEWS.MAJOR_VIEW) && getNglView(VIEWS.MAJOR_VIEW).stage;
@@ -310,7 +310,6 @@ const MoleculeView = memo(
       useSelector(state => state.previewReducers.molecule.disableNglControlButtons[currentID]) || {};
 
     let tagEditIconVisible = true;
-    // const disableUserInteraction = useDisableUserInteraction();
 
     const oldUrl = useRef('');
     const setOldUrl = url => {
