+++ js/components/preview/moleculeGroups/molGroupSelector.js
@@ -9,7 +9,6 @@ import { useDispatch } from 'react-redux';
 import { VIEWS } from '../../../constants/constants';
 import { withLoadingMolGroupList } from './withLoadingMolGroupList';
 import { NglContext } from '../../nglView/nglProvider';
-// import { useDisableUserInteraction } from '../../helpers/useEnableUserInteracion';
 import { clearMoleculeGroupSelection } from './redux/dispatchActions';
 
 export const heightOfBody = '164px';
@@ -34,7 +33,6 @@ const MolGroupSelector = memo(({ handleHeightChange }) => {
   const ref = useRef(null);
 
   const { getNglView } = useContext(NglContext);
-  // const disableUserInteraction = useDisableUserInteraction();
   const dispatch = useDispatch();
 
   return (
