+++ js/components/preview/tags/details/tagDetails.js
@@ -83,7 +83,7 @@ const TagDetails = memo(({ handleHeightChange }) => {
   const [elementHeight, setElementHeight] = useState(0);
   const [sortSwitch, setSortSwitch] = useState(0);
 
-  const preTagList = useSelector(state => state.selectionReducers.tagList);
+  const preTagList = useSelector(state => state.apiReducers.tagList);
   const [tagList, setTagList] = useState([]);
 
   const [searchString, setSearchString] = useState(null);
