+++ js/components/preview/tags/redux/dispatchActions.js
@@ -1,17 +1,8 @@
 import {
   setSelectedTagList,
   appendSelectedTagList,
-  removeFromSelectedTagList,
-  setCategoryList,
-  setTagList,
-  appendTagList,
-  appendToDisplayAllNGLList,
-  removeFromDisplayAllNGLList,
-  appendToListAllForTagList,
-  removeFromListAllForTagList
+  removeFromSelectedTagList
 } from '../../../../reducers/selection/actions';
-import { CATEGORY_TYPE } from '../../../../constants/constants';
-import { addLigand, removeLigand } from '../../molecule/redux/dispatchActions';
 import {
   setProteinList,
   setDensityList,
@@ -23,20 +14,20 @@ import {
   setFragmentDisplayList,
   setMolGroupSelection,
   setVectorList,
-  setVectorOnList,
-  updateTag
+  setVectorOnList
 } from '../../../../reducers/selection/actions';
 import {
   setMolGroupOn,
   updateMoleculeTag,
   setAllMolLists,
-  setMoleculeTags,
-  setDownloadTags,
-  setNoTagsReceived
+  setNoTagsReceived,
+  updateTag,
+  setTagList,
+  appendTagList,
+  setCategoryList
 } from '../../../../reducers/api/actions';
 import { setSortDialogOpen } from '../../molecule/redux/actions';
 import { resetCurrentCompoundsSettings } from '../../compounds/redux/actions';
-import { getRandomColor } from '../../molecule/utils/color';
 import { updateExistingTag, getAllData } from '../api/tagsApi';
 import {
   getMoleculeTagForTag,
@@ -46,7 +37,6 @@ import {
   getCategoryIds
 } from '../utils/tagUtils';
 import { DJANGO_CONTEXT } from '../../../../utils/djangoContext';
-import { diffBetweenDatesInDays } from '../../../../utils/common';
 
 export const setTagSelectorData = (categories, tags) => dispatch => {
   dispatch(setCategoryList(categories));
@@ -63,13 +53,13 @@ export const removeSelectedTag = tagItem => dispatch => {
 
 export const selectAllTags = () => (dispatch, getState) => {
   const state = getState();
-  let tagList = state.selectionReducers.tagList;
+  let tagList = state.apiReducers.tagList;
   tagList.forEach(t => dispatch(appendSelectedTagList(t)));
 };
 
 export const clearAllTags = () => (dispatch, getState) => {
   const state = getState();
-  let tagList = state.selectionReducers.tagList;
+  let tagList = state.apiReducers.tagList;
   tagList.forEach(t => dispatch(removeFromSelectedTagList(t)));
 };
 
@@ -126,35 +116,6 @@ export const storeData = data => (dispatch, getState) => {
 
   let allMolecules = [];
   data.molecules.forEach(mol => {});
-};
-
-export const displayAllMolsInNGL = (stage, tag) => (dispatch, getState) => {
-  const state = getState();
-  const allMolecules = state.apiReducers.all_mol_lists;
-  const taggedMolecules = allMolecules.filter(mol => mol.tags_set.some(id => id === tag.id));
-  taggedMolecules.forEach(mol => {
-    const color = getRandomColor(mol);
-    dispatch(addLigand(stage, mol, color, false, true, true));
-  });
-  dispatch(appendToDisplayAllNGLList(tag));
-};
-
-export const hideAllMolsInNGL = (stage, tag) => (dispatch, getState) => {
-  const state = getState();
-  const allMolecules = state.apiReducers.all_mol_lists;
-  const taggedMolecules = allMolecules.filter(mol => mol.tags_set.some(id => id === tag.id));
-  taggedMolecules.forEach(mol => {
-    dispatch(removeLigand(stage, mol, true));
-  });
-  dispatch(removeFromDisplayAllNGLList(tag));
-};
-
-export const displayInListForTag = tag => dispatch => {
-  dispatch(appendToListAllForTagList(tag));
-};
-
-export const hideInListForTag = tag => dispatch => {
-  dispatch(removeFromListAllForTagList(tag));
 };
 
 export const updateTagProp = (tag, value, prop) => (dispatch, getState) => {
