+++ js/components/preview/tags/tagCategory.js
@@ -18,8 +18,8 @@ const useStyles = makeStyles(theme => ({
 const TagCategory = memo(({ headerPadding = 0, tagClickCallback, disabled = false }) => {
   const classes = useStyles();
 
-  const categoryList = useSelector(state => state.selectionReducers.categoryList);
-  let tagList = useSelector(state => state.selectionReducers.tagList);
+  const categoryList = useSelector(state => state.apiReducers.categoryList);
+  let tagList = useSelector(state => state.apiReducers.tagList);
   tagList = tagList.sort(compareTagsAsc);
 
   const siteCategory = categoryList.find(c => c.category === CATEGORY_TYPE.SITE);
