+++ js/reducers/api/constants.js
@@ -40,5 +40,12 @@ export const constants = {
   SET_DOWNLOAD_TAGS: prefix + 'SET_DOWNLOAD_TAGS',
   APPEND_TO_DOWNLOAD_TAGS: prefix + 'APPEND_TO_DOWNLOAD_TAGS',
   SET_DIRECT_DOWNLOAD_IN_PROGRESS: prefix + 'SET_DIRECT_DOWNLOAD_IN_PROGRESS',
-  SET_SNAPSHOT_DOWNLOAD_URL: prefix + 'SET_SNAPSHOT_DOWNLOAD_URL'
+  SET_SNAPSHOT_DOWNLOAD_URL: prefix + 'SET_SNAPSHOT_DOWNLOAD_URL',
+
+  SET_TAG_LIST: prefix + 'SET_TAG_LIST',
+  UPDATE_TAG: prefix + 'UPDATE_TAG',
+  APPEND_TAG_LIST: prefix + 'APPEND_TAG_LIST',
+  REMOVE_FROM_TAG_LIST: prefix + 'REMOVE_FROM_TAG_LIST',
+
+  SET_CATEGORY_LIST: prefix + 'SET_CATEGORY_LIST'
 };
