+++ js/reducers/selection/constants.js
@@ -1,8 +1,6 @@
 const prefix = 'REDUCERS_SELECTION_';
 
 export const constants = {
-  INCREMENT_COUNT_OF_PENDING_VECTOR_LOAD_REQUESTS: prefix + 'INCREMENT_COUNT_OF_PENDING_VECTOR_LOAD_REQUESTS',
-  DECREMENT_COUNT_OF_PENDING_VECTOR_LOAD_REQUESTS: prefix + 'DECREMENT_COUNT_OF_PENDING_VECTOR_LOAD_REQUESTS',
   SET_OBJECT_SELECTION: prefix + 'SET_OBJECT_SELECTION',
 
   SET_TO_BUY_LIST: prefix + 'SET_TO_BUY_LIST',
@@ -35,9 +33,6 @@ export const constants = {
   APPEND_QUALITY_LIST: prefix + 'APPEND_QUALITY_LIST',
   REMOVE_FROM_QUALITY_LIST: prefix + 'REMOVE_FROM_QUALITY_LIST',
 
-  SET_LIST_ALL_FOR_TAG_LIST: prefix + 'SET_LIST_ALL_FOR_TAG_LIST',
-  APPEND_TO_LIST_ALL_FOR_TAG_LIST: prefix + 'APPEND_TO_LIST_ALL_FOR_TAG_LIST',
-  REMOVE_FROM_LIST_ALL_FOR_TAG_LIST: prefix + 'REMOVE_FROM_LIST_ALL_FOR_TAG_LIST',
   SET_DISPLAY_ALL_NGL_LIST: prefix + 'SET_DISPLAY_ALL_NGL_LIST',
   APPEND_TO_DISPLAY_ALL_NGL_LIST: prefix + 'APPEND_TO_DISPLAY_ALL_NGL_LIST',
   REMOVE_FROM_DISPLAY_ALL_NGL_LIST: prefix + 'REMOVE_FROM_DISPLAY_ALL_NGL_LIST',
@@ -60,13 +55,6 @@ export const constants = {
   APPEND_DENSITY_TYPE: prefix + 'APPEND_DENSITY_TYPE',
   REMOVE_DENSITY_TYPE: prefix + 'REMOVE_DENSITY_TYPE',
 
-  SET_CATEGORY_LIST: prefix + 'SET_CATEGORY_LIST',
-  APPEND_CATEGORY_LIST: prefix + 'APPEND_CATEGORY_LIST',
-  REMOVE_FROM_CATEGORY_LIST: prefix + 'REMOVE_FROM_CATEGORY_LIST',
-  SET_TAG_LIST: prefix + 'SET_TAG_LIST',
-  UPDATE_TAG: prefix + 'UPDATE_TAG',
-  APPEND_TAG_LIST: prefix + 'APPEND_TAG_LIST',
-  REMOVE_FROM_TAG_LIST: prefix + 'REMOVE_FROM_TAG_LIST',
   SET_SELECTED_TAG_LIST: prefix + 'SET_SELECTED_TAG_LIST',
   APPEND_SELECTED_TAG_LIST: prefix + 'APPEND_SELECTED_TAG_LIST',
   REMOVE_FROM_SELECTED_TAG_LIST: prefix + 'REMOVE_FROM_SELECTED_TAG_LIST',
@@ -78,7 +66,6 @@ export const constants = {
   UPDATE_VECTOR_COMPOUNDS: prefix + 'UPDATE_VECTOR_COMPOUNDS',
   RESET_BOND_COLOR_MAP_OF_VECTORS: prefix + 'RESET_BOND_COLOR_MAP_OF_VECTORS',
   UPDATE_BOND_COLOR_MAP_OF_COMPOUNDS: prefix + 'UPDATE_BOND_COLOR_MAP_OF_COMPOUNDS',
-  SET_DISPLAYED_MOLECULES_HIT_NAV: prefix + 'SET_DISPLAYED_MOLECULES_HIT_NAV',
   SET_IS_TAG_GLOBAL_EDIT: prefix + 'SET_IS_TAG_GLOBAL_EDIT',
   SET_MOL_LIST_TO_EDIT: prefix + 'SET_MOL_LIST_TO_EDIT',
   APPEND_TO_MOL_LIST_TO_EDIT: prefix + 'APPEND_TO_MOL_LIST_TO_EDIT',
@@ -87,12 +74,9 @@ export const constants = {
   SET_TAG_TO_EDIT: prefix + 'SET_TAG_TO_EDIT',
   SET_DISPLAY_ALL_MOLECULES: prefix + 'SET_DISPLAY_ALL_MOLECULES',
   SET_DISPLAY_UNTAGGED_MOLECULES: prefix + 'SET_DISPLAY_UNTAGGED_MOLECULES',
-  SET_ASSOCIATED_DOWNLOAD_TAG_NAME: prefix + 'SET_ASSOCIATED_DOWNLOAD_TAG_NAME',
   SET_SELECT_ALL_MOLECULES: prefix + 'SET_SELECT_ALL_MOLECULES',
   SET_UNSELECT_ALL_MOLECULES: prefix + 'SET_UNSELECT_ALL_MOLECULES',
-  SET_NEXT_X_MOLECULES: prefix + 'SET_NEXT_X_MOLECULES',
-
-  RESET_SELECTION_STATE_ON_SNAPSHOT_CHANGE: prefix + 'RESET_SELECTION_STATE_ON_SNAPSHOT_CHANGE'
+  SET_NEXT_X_MOLECULES: prefix + 'SET_NEXT_X_MOLECULES'
 };
 
 export const PREDEFINED_FILTERS = {
