+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fragalysis-frontend",
-  "version": "0.10.92",
+  "version": "0.10.93",
   "description": "Frontend for fragalysis",
   "main": "webpack.config.js",
   "scripts": {
